02-Master_2/02-Knative-REST/requirements.txt
parliament-functions==0.1.0
mysql.connector
datetime
